+++ app/src/main/java/org/keepgoeat/presentation/my/MyViewModel.kt
@@ -8,7 +8,6 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import org.keepgoeat.data.datasource.local.KGEDataSource
 import org.keepgoeat.domain.model.AchievedGoal
-import org.keepgoeat.domain.repository.AuthRepository
 import org.keepgoeat.domain.repository.GoalRepository
 import org.keepgoeat.presentation.type.SortType
 import org.keepgoeat.util.UiState
@@ -16,19 +15,16 @@ import javax.inject.Inject
 
 @HiltViewModel
 class MyViewModel @Inject constructor(
-    private val authRepository: AuthRepository,
     private val goalRepository: GoalRepository,
     private val localStorage: KGEDataSource,
 ) : ViewModel() {
-    private val _achievedGoalUiState = MutableStateFlow<UiState<List<AchievedGoal>>>(UiState.Loading)
+    private val _achievedGoalUiState =
+        MutableStateFlow<UiState<List<AchievedGoal>>>(UiState.Loading)
     val achievedGoalUiState get() = _achievedGoalUiState.asStateFlow()
     private val _logoutUiState = MutableStateFlow<UiState<Boolean>>(UiState.Loading)
     val logoutUiState get() = _logoutUiState.asStateFlow()
     private val _achievedGoalCount = MutableStateFlow(0)
     val achievedGoalCount get() = _achievedGoalCount.asStateFlow()
-    private val _deleteAccountUiState =
-        MutableStateFlow<UiState<Boolean>>(UiState.Loading)
-    val deleteAccountUiState get() = _deleteAccountUiState.asStateFlow()
     val loginPlatForm = localStorage.loginPlatform
     val userName = localStorage.userName
     val userEmail = localStorage.userEmail
@@ -52,16 +48,5 @@ class MyViewModel @Inject constructor(
     fun logout() {
         localStorage.clear()
         _logoutUiState.value = UiState.Success(true)
-    }
-
-    fun deleteAccount() {
-        viewModelScope.launch {
-            authRepository.deleteAccount()
-                .onSuccess {
-                    _deleteAccountUiState.value = UiState.Success(true)
-                }.onFailure {
-                    _deleteAccountUiState.value = UiState.Error(it.message)
-                }
-        }
     }
 }
